+++ src/classes/task-edit.js
@@ -1,14 +1,16 @@
-import {createContainer} from "../utils";
+import {Component} from "./component";
 
-export class TaskEdit {
+export class TaskEdit extends Component {
   constructor(data) {
+    super();
     this._title = data.title;
     this._dueDate = data.dueDate;
     this._tags = data.tags;
     this._picture = data.picture;
     this._repeatingDays = data.repeatingDays;
 
-    this._element = null;
+    this._onSubmitButtonClick = this._onSubmitButtonClick.bind(this);
+
     this._onSubmit = null;
   }
 
@@ -23,10 +25,6 @@ export class TaskEdit {
 
   set onSubmit(fn) {
     this._onSubmit = fn;
-  }
-
-  get element() {
-    return this._element;
   }
 
   get template() {
@@ -150,24 +148,14 @@ export class TaskEdit {
     </article>`.trim();
   }
 
-  render() {
-    this._element = createContainer(this.template);
-    this.bind();
-    return this._element;
-  }
-
-  unrender() {
-    this.unbind();
-    this._element = null;
-  }
-
   bind() {
     this._element.querySelector(`.card__form`)
-      .addEventListener(`submit`, this._onSubmitButtonClick.bind(this));
+      .addEventListener(`submit`, this._onSubmitButtonClick);
   }
 
   unbind() {
-    // Удаление обработчиков
+    this._element.querySelector(`.card__form`)
+      .removeEventListener(`submit`, this._onSubmitButtonClick);
   }
 
 }
