+++ src/utils.js
@@ -20,7 +20,7 @@ export const formatAMPM = (date) => {
   return strTime;
 };
 
-export const createContainer = (template) => {
+export const createElement = (template) => {
   const newElement = document.createElement(`div`);
   newElement.innerHTML = template;
   return newElement.firstChild;
